[CHATBOT-630] Package reorganization.
