+++ vertigo-bot-designer/src/main/java/io/vertigo/chatbot/designer/admin/services/MonitoringServices.java
@@ -16,8 +16,8 @@ import io.vertigo.chatbot.commons.domain.ChatbotNode;
 import io.vertigo.chatbot.commons.domain.RunnerHealthCheck;
 import io.vertigo.chatbot.designer.builder.monitoring.MonitoringPAO;
 import io.vertigo.chatbot.designer.builder.services.NodeServices;
-import io.vertigo.chatbot.designer.builder.services.TrainingServices;
 import io.vertigo.chatbot.designer.builder.services.bot.ChatbotServices;
+import io.vertigo.chatbot.designer.builder.services.training.TrainingServices;
 import io.vertigo.chatbot.designer.domain.commons.Person;
 import io.vertigo.chatbot.designer.domain.monitoring.AlertingEvent;
 import io.vertigo.chatbot.designer.domain.monitoring.MonitoringAlertingSubscription;
