+++ vertigo-bot-designer/src/main/java/io/vertigo/chatbot/designer/builder/controllers/bot/ModelListController.java
@@ -37,10 +37,10 @@ import io.vertigo.chatbot.commons.domain.SavedTrainingCriteria;
 import io.vertigo.chatbot.commons.domain.TrainerInfo;
 import io.vertigo.chatbot.commons.domain.Training;
 import io.vertigo.chatbot.designer.builder.services.NodeServices;
-import io.vertigo.chatbot.designer.builder.services.SavedTrainingServices;
-import io.vertigo.chatbot.designer.builder.services.TrainerInfoServices;
-import io.vertigo.chatbot.designer.builder.services.TrainingServices;
 import io.vertigo.chatbot.designer.builder.services.bot.ChatbotCustomConfigServices;
+import io.vertigo.chatbot.designer.builder.services.training.SavedTrainingServices;
+import io.vertigo.chatbot.designer.builder.services.training.TrainerInfoServices;
+import io.vertigo.chatbot.designer.builder.services.training.TrainingServices;
 import io.vertigo.ui.core.ViewContext;
 import io.vertigo.ui.core.ViewContextKey;
 import io.vertigo.ui.impl.springmvc.argumentresolvers.ViewAttribute;
